Create a session via the Client
